requirements.txt
numpy>=1.26.4
torch>=2.3.1
torchvision>=0.18.1
Pillow>=11.3.0
pandas>=2.3.3
matplotlib>=3.10.7
opencv-python>=4.11.0
scipy>=1.15.3
omegaconf>=2.3.0
tqdm>=4.67.1
imageio>=2.37.2
rootutils>=1.0.7
hydra-core>=1.3.2

# Optional: needed for distance-based pose evaluation (eval_dist.py)
# evo>=1.11.0
